refactor(generator): 抽象数据库表结构相关接口并替换实现

+++ mybatisx-plugin/src/main/java/com/baomidou/mybatisx/feat/mybatis/generator/plugin/IntellijColumnInfo.java
@@ -15,5 +15,5 @@ public class IntellijColumnInfo {
   private String remarks;
   private String columnDefaultValue;
   private Boolean nullable;
-  private short keySeq;
+  private int keySeq;
 }
